Events Object Class.
Adds a new FormPanel and puts two line
itmes (Label and Field) together with an OKButton. Introduces new class
FormEvent which extends EventObject class and uses this class to store
the information from the Fields on the FormPanel. This custom event is
raised when the okay button in clicked by following the system
ActionListener for the button .. It then uses a custom
listener to send the event object back to the MainFrame, which calls the
method in the MainPanel to append the text to the TextArea of the
MainPanel.