Replaces Strings in ListBox with AgeCat class
Creates a simple utility class AgeCat for matching up IDs with Strings.
This class is then added to the ListBox in place of just strings. The
AgeCat class also has two methods to return either just the ID or the
String itself. Yet to be implemented into the FormEvent class. See next
commit.